Move animation scheduler to scheduler directory.
